requirements.txt
numpy==1.24.3
opencv-python>=4.5.0
Pillow>=8.0.0
pyautogui>=0.9.53
keyboard>=0.13.5
mouse>=0.7.1
mss>=6.1.0
pywin32>=300
torch==2.3.0+cu121
torchvision==0.18.0+cu121
--extra-index-url https://download.pytorch.org/whl/cu121
matplotlib>=3.4.0
scikit-learn>=0.24.0
segment-anything>=1.0
clip @ git+https://github.com/openai/CLIP.git
PyYAML>=5.4.0
openai>=1.0.0
anthropic>=0.8.0
requests>=2.31.0
wandb>=0.19.9
python-dotenv
flask==3.0.3
dash==3.0.3
dash_cytoscape==1.0.2
langgraph 
langchain
langchain-openai
pytesseract